+++ internal/lspserver/server.go
@@ -1243,7 +1243,7 @@ func parameterLabel(param intel.Parameter) string {
 }
 
 func completionTriggerCharacters() []string {
-	return []string{".", "\"", "'"}
+	return []string{".", "\"", "'", "@"}
 }
 
 func max(a, b int) int {